Start the process of breaking e2e tests into their own modules/functions
