+++ tests/e2e_tests/registration_tests.rs
@@ -1,0 +1,3 @@
+pub fn test_registration() -> Result<(), Box<dyn std::error::Error>> {
+    Ok(())
+}
